imap: add mailbox stub
